Setup release workflow
Ensure the example project can be published to Bintray.

Unfortunately bintrayRelease task on the root of the project needs to be stubbed out as it incorrectly publishes the release despite root of the project not having any artifact to publish.

This also means the examples module will need to have the newly cut Bintray release published.

SBT is upgraded to 1.1.0 due to issues around publish settings:

https://github.com/sbt/sbt/issues/3136#issuecomment-296817619